requirements.txt
kaleido==0.2.1
python-dotenv==1.0.0
plotly==5.15.0